+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## Change log
 ----------------------
 
+Version 3.4
+-------------
+
+ADDED:
+
+- new method that resolves if the given class is an primitive array type
+
+CHANGED:
+
+- migrate to gradle build system
+- update of test-dependency silly-collections version to 5.8
+- removed deprecated methods in ClassExtensions class
+- removed deprecated methods in ReflectionExtensions class
+- method newInstance from the ReflectionExtensions class considers now array types on creation
+
 Version 3.3
 -------------